+++ src/main/java/its_meow/betteranimalsplus/common/entity/util/abstracts/EntityWaterMobPathingWithTypesBucketable.java
@@ -65,7 +65,7 @@ public abstract class EntityWaterMobPathingWithTypesBucketable extends EntityWat
     @Override
     public void livingTick() {
         if(!this.isInWater() && this.onGround && this.collidedVertically) {
-            this.setMotion(this.getMotion().add((double) ((this.rand.nextFloat() * 2.0F - 1.0F) * 0.05F), (double) 0.4F, (double) ((this.rand.nextFloat() * 2.0F - 1.0F) * 0.05F)));
+            this.setMotion(this.getMotion().add((this.rand.nextFloat() * 2.0F - 1.0F) * 0.05F, 0.4F, (this.rand.nextFloat() * 2.0F - 1.0F) * 0.05F));
             this.onGround = false;
             this.isAirBorne = true;
             this.playSound(this.getFlopSound(), this.getSoundVolume(), this.getSoundPitch());
